test: add scroll to top
